move network state files to separate folder
