+++ test/modules/header/components/Header.test.js
@@ -6,8 +6,8 @@ import { TestUtils } from '../../../test-utils.js';
 import { $injector } from '../../../../src/injection';
 import { OlCoordinateService } from '../../../../src/services/OlCoordinateService';
 import { layersReducer } from '../../../../src/modules/map/store/layers.reducer';
-import { networkReducer } from '../../../../src/store/network.reducer';
-import { setFetching } from '../../../../src/store/network.action';
+import { networkReducer } from '../../../../src/store/network/network.reducer';
+import { setFetching } from '../../../../src/store/network/network.action';
 import { MainMenuTabIndex } from '../../../../src/modules/menu/components/mainMenu/MainMenu';
 
 window.customElements.define(Header.tag, Header);
