+++ test/store/network.reducer.test.js
@@ -1,5 +1,5 @@
-import { networkReducer } from '../../src/store/network.reducer';
-import { setFetching, setOffline } from '../../src/store/network.action';
+import { networkReducer } from '../../src/store/network/network.reducer';
+import { setFetching, setOffline } from '../../src/store/network/network.action';
 import { TestUtils } from '../test-utils.js';
